requirements.txt
asttokens==2.0.5
backcall==0.2.0
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
decorator==5.1.1
executing==0.8.2
fonttools==4.29.1
html5lib==1.1
importlib-metadata==4.11.1
ipython==8.1.0
jedi==0.18.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.1.2
Pygments==2.11.2
pyphen==0.12.0
python-frontmatter==1.0.0
PyYAML==5.1
six==1.16.0
stack-data==0.2.0
tinycss2==1.1.1
traitlets==5.1.1
wcwidth==0.2.5
weasyprint==54.1
webencodings==0.5.1
zipp==3.7.0
zopfli==0.1.9
